Unnecessary animation frame!
Aha! There was an odd animation overlap!

section div#rain::before {
  content: "";
  position: absolute;
  width: 100%;
  height: 100%;
  background-image: url(Images/lightningBright.png);
  animation: lightning 0.4s linear infinite;
  opacity: 0;
}

Since the lightning animation is linked to the lightning keyframe- we didn't need the ::before rain animation.
Thank you for looking at this for me! It has made me more confident that I can add extra layers if I need to-- and it's so tidy! That's something I need to work on!